Started on flows module
Interrupted as I consider how to handle mounts - right now the easiest
way forward is to accept a mount configuration the same way that the
executions.Execute method does, but this needs to be addressed today or
tomorrow.

Also thought a lot about the CLI hierarchy. It makes more sense as

```
simplex:
    components:
        <CRUD>
        builds:
            <CRUD>
        executions:
            <CRUD>
    flows:
        <CRUD>
        builds:
            <CRUD>
        executions:
            <CRUD>
```

There will be an appropriate symmetry between flows and components.