flows: validate -> materialize specifications
This is a serious change and is a work in progress.

Next, we are going to strip out the `--mounts` option from `shn flows
execute` in favor of specifying mounts in the flow specification.

We also have to modify the flows Execute method to use specification
structs with the new members (Mounts and Stages).