Flow execution now accepts environment variables
Set on a per-step basis. This required the components.Execute method
signature to be changed to accept a map[string]string argument with
environment variable overrides.

The materialization of the flow specification manually runs the
materialization loop (just as in components materialization of run
specification). It is okay for now, but if this logic keeps repeating it
may make sense to add a function exposing the same functionality in the
components package.

The CLI does not accept an environment variable override from the
command line when executing a component. It doesn't really make sense
for it to do so.

Resolves https://github.com/simiotics/shnorky/issues/22